chore: renames report-web-vitals
