+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react"
 import ReactDOM from "react-dom"
 import "./index.css"
 import Glimpse from "./App.js"
-import reportWebVitals from "./reportWebVitals"
+import reportWebVitals from "./report-web-vitals.js"
 import "font-awesome/css/font-awesome.min.css"
 
 ReactDOM.render(
